# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Game API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/endgame": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "End Game",
                "description": "Players End Game",
                "operationId": "endGame",
                "requestBody": {
                    "description": "Players End Game",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Players": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "username",
                                                    "example": "3"
                                                },
                                                "score": {
                                                    "type": "string",
                                                    "format": "password",
                                                    "example": "10"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Player not found"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/leaderboard": {
            "get": {
                "tags": [
                    "Leaderboard"
                ],
                "summary": "Leaderboard",
                "description": "Leaderboard",
                "operationId": "leaderboard",
                "responses": {
                    "200": {
                        "description": "Leaderboard"
                    }
                }
            }
        },
        "/api/v1/user/signin": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Login by username, password",
                "operationId": "signIn",
                "requestBody": {
                    "description": "User Login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "username",
                                        "example": "haldun"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "qwq123qwq"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The password must be at least 5 characters."
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Sign Up",
                "description": "Sign up by username, password",
                "operationId": "signUp",
                "requestBody": {
                    "description": "User Sign Up",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "username",
                                        "example": "haldun"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "qwq123qwq"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "409": {
                        "description": "User allready exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "User allready exist"
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}